Moved googleLogin into its own file
